+++ models/marts/core/dim_section.sql
@@ -1,20 +1,5 @@
 
-WITH primary_teacher AS (
-
-
-    SELECT
-        ssa.section_reference.school_year,
-        ssa.section_reference.school_id,
-        ssa.section_reference.session_name,
-        ssa.section_reference.local_course_code,
-        ssa.staff_reference.staff_unique_id
-    FROM {{ ref('stg_edfi_staff_section_associations') }} ssa
-    WHERE classroom_position_descriptor = 'Teacher of Record'
-
-)
-
-
-SELECT
+SELECT DISTINCT
     {{ dbt_utils.surrogate_key([
         'sections.course_offering_reference.school_id',
         'sections.course_offering_reference.school_year',
@@ -35,10 +20,6 @@ SELECT
         'sections.course_offering_reference.school_year',
         'sections.course_offering_reference.session_name'
     ]) }}                                                       AS session_key,
-    {{ dbt_utils.surrogate_key([
-        'primary_teacher.staff_unique_id',
-        'primary_teacher.school_year'
-     ]) }}                                                      AS primary_staff_key,
     course_offerings.session_reference.school_year              AS school_year,
     sections.section_identifier                                 AS section_identifier,
     COALESCE(
@@ -63,11 +44,6 @@ LEFT JOIN {{ ref('stg_edfi_course_offerings') }} course_offerings
     AND course_offerings.school_reference.school_id = sections.course_offering_reference.school_id
     AND course_offerings.session_reference.school_year = sections.course_offering_reference.school_year
     AND course_offerings.session_reference.session_name = sections.course_offering_reference.session_name
-LEFT JOIN primary_teacher
-    ON course_offerings.local_course_code = primary_teacher.local_course_code
-    AND course_offerings.school_reference.school_id = primary_teacher.school_id
-    AND course_offerings.session_reference.school_year = primary_teacher.school_year
-    AND course_offerings.session_reference.session_name = primary_teacher.session_name
 LEFT JOIN {{ ref('stg_edfi_courses') }} courses
     ON course_offerings.school_year = courses.school_year
     AND courses.course_code = course_offerings.course_reference.course_code